removed [id] api and modified dairy-component page
